chore(nodejs): move ingestion producer registry out of common
Relocate ingestion-specific code that lived in the shared common layer so
that common no longer imports ingestion.

- Move the ingestion producer registry (createIngestionProducerRegistry,
  KafkaProducerRegistryComponent, UPSTREAM/DOWNSTREAM slot names, config
  maps, env configs) from common/outputs and ingestion/common/config into a
  new ingestion/common/producers.ts, mirroring cdp/outputs/producers.ts.
- Move PersonHogConfig into common/personhog, next to createPersonHogClient.
- Replace redis-pools' IngestionConsumerConfig dependency with a local
  CookielessRedisConfig type.

The generic producer registry machinery (KafkaProducerRegistry,
KafkaProducerRegistryBuilder) stays in common/outputs. common/outputs,
common/personhog and config/redis-pools no longer import ingestion.
