+++ nodejs/src/common/personhog/index.ts
@@ -1,6 +1,6 @@
+import type { CommonConfig } from '~/common/config'
 import { GroupRepository } from '~/common/groups/repositories/group-repository.interface'
 import { PersonRepository } from '~/common/persons/repositories/person-repository'
-import { PersonHogConfig } from '~/ingestion/config'
 import { logger } from '~/utils/logger'
 
 import { PersonHogClient, parseRolloutTeamIds } from './client'
@@ -11,6 +11,27 @@ export { PersonHogClient } from './client'
 export type { PersonHogClientConfig } from './client'
 export { PersonHogGroupRepository } from './personhog-group-repository'
 export { PersonHogPersonRepository } from './personhog-person-repository'
+
+/** PersonHog gRPC client config */
+export type PersonHogConfig = Pick<
+    CommonConfig,
+    | 'PERSONHOG_ENABLED'
+    | 'PERSONHOG_ADDR'
+    | 'PERSONHOG_GROUPS_ROLLOUT_PERCENTAGE'
+    | 'PERSONHOG_GROUPS_ROLLOUT_TEAM_IDS'
+    | 'PERSONHOG_PERSONS_ROLLOUT_PERCENTAGE'
+    | 'PERSONHOG_PERSONS_ROLLOUT_TEAM_IDS'
+    | 'PERSONHOG_TLS'
+    | 'PERSONHOG_TIMEOUT_MS'
+    | 'PERSONHOG_READ_MAX_BYTES'
+    | 'PERSONHOG_WRITE_MAX_BYTES'
+    | 'PERSONHOG_PING_INTERVAL_MS'
+    | 'PERSONHOG_PING_TIMEOUT_MS'
+    | 'PERSONHOG_PING_IDLE_CONNECTION'
+    | 'PERSONHOG_IDLE_CONNECTION_TIMEOUT_MS'
+    | 'PERSONHOG_STATE_MONITOR_POLL_INTERVAL_MS'
+    | 'PLUGIN_SERVER_MODE'
+>
 
 export function createPersonHogClient(config: PersonHogConfig): PersonHogClient | null {
     if (!config.PERSONHOG_ENABLED || !config.PERSONHOG_ADDR) {
