+++ nodejs/src/config/redis-pools.ts
@@ -1,5 +1,4 @@
 import { CommonConfig } from '../common/config'
-import { IngestionConsumerConfig } from '../ingestion/config'
 import { RedisConnectionConfig } from '../utils/db/redis'
 
 /**
@@ -51,6 +50,18 @@ export function createPosthogRedisConnectionConfig(
     return { url: config.REDIS_URL, name: 'posthog-redis' }
 }
 
+/** Cookieless Redis connection params — COOKIELESS_REDIS_* are ingestion-specific config keys. */
+export type CookielessRedisConfig = {
+    COOKIELESS_REDIS_HOST: string
+    COOKIELESS_REDIS_PORT: number
+}
+
+/** $feature_flag_called dedup Redis connection params — these are ingestion-specific config keys. */
+export type FeatureFlagCalledDedupRedisConfig = {
+    INGESTION_FEATURE_FLAG_CALLED_DEDUP_REDIS_HOST: string
+    INGESTION_FEATURE_FLAG_CALLED_DEDUP_REDIS_PORT: number
+}
+
 /**
  * Build the connection config for the $feature_flag_called dedup Redis pool.
  *
@@ -59,10 +70,7 @@ export function createPosthogRedisConnectionConfig(
  * INGESTION_FEATURE_FLAG_CALLED_DEDUP_REDIS_HOST points at the dedicated instance.
  */
 export function createFeatureFlagCalledDedupRedisConnectionConfig(
-    config: Pick<
-        IngestionConsumerConfig,
-        'INGESTION_FEATURE_FLAG_CALLED_DEDUP_REDIS_HOST' | 'INGESTION_FEATURE_FLAG_CALLED_DEDUP_REDIS_PORT'
-    > &
+    config: FeatureFlagCalledDedupRedisConfig &
         Pick<
             CommonConfig,
             | 'INGESTION_REDIS_HOST'
@@ -88,8 +96,7 @@ export function createFeatureFlagCalledDedupRedisConnectionConfig(
  * Fallback chain: COOKIELESS_REDIS_HOST → REDIS_URL
  */
 export function createCookielessRedisConnectionConfig(
-    config: Pick<IngestionConsumerConfig, 'COOKIELESS_REDIS_HOST' | 'COOKIELESS_REDIS_PORT'> &
-        Pick<CommonConfig, 'REDIS_URL'>
+    config: CookielessRedisConfig & Pick<CommonConfig, 'REDIS_URL'>
 ): RedisConnectionConfig {
     if (config.COOKIELESS_REDIS_HOST) {
         return {
