+++ nodejs/src/ingestion/common/producers.ts
@@ -1,8 +1,31 @@
 import type { AllowedConfigKey } from '~/common/outputs/kafka-producer-config'
 
+// Producer names
+//
+// A producer name is a named Kafka connection slot, not a fixed cluster: the code declares the
+// slots, and each pipeline's Helm charts wire a slot to a concrete cluster (broker list +
+// security protocol) and route its outputs to it. A slot can map to different clusters in
+// different pipelines. The slot → cluster mapping lives in the charts repo
+// (shared/ingestion/common*.yaml, argocd/ingestion/config/*.yaml).
+//
+// Cluster-accurate slots:
+//   INGESTION_UPSTREAM   — dedicated ingestion cluster; re-consumed topics (overflow/async/dlq)
+//   INGESTION_DOWNSTREAM — warpstream-ingestion cluster; ClickHouse-bound outputs
+// Pipeline-specific slots (e.g. session replay's warpstream-replay producer) live in their own module.
+
+/** UPSTREAM — dedicated ingestion cluster; re-consumed topics (overflow/async/dlq). */
+export const INGESTION_UPSTREAM_PRODUCER = 'INGESTION_UPSTREAM' as const
+export type IngestionUpstreamProducer = typeof INGESTION_UPSTREAM_PRODUCER
+
+/** DOWNSTREAM — warpstream-ingestion cluster; ClickHouse-bound outputs. */
+export const INGESTION_DOWNSTREAM_PRODUCER = 'INGESTION_DOWNSTREAM' as const
+export type IngestionDownstreamProducer = typeof INGESTION_DOWNSTREAM_PRODUCER
+
+/** Union of all known producer names. Extend this as new producers are added. */
+export type ProducerName = IngestionUpstreamProducer | IngestionDownstreamProducer
+
 // =============================================================================
-// Cluster-accurate producer slots: INGESTION_UPSTREAM / INGESTION_DOWNSTREAM
-// (see common/outputs/index.ts for the slot contract).
+// Cluster-accurate producer slots: env-var maps and their default env config.
 // =============================================================================
 
 /** UPSTREAM — dedicated ingestion cluster (re-consumed topics); ssl + sasl auth. */
