+++ nodejs/src/ingestion/pipelines/errortracking/config.ts
@@ -1,4 +1,3 @@
-import { INGESTION_DOWNSTREAM_PRODUCER, INGESTION_UPSTREAM_PRODUCER, type ProducerName } from '~/common/outputs'
 import {
     KAFKA_APP_METRICS_2,
     KAFKA_CLICKHOUSE_TOPHOG,
@@ -9,6 +8,11 @@ import {
     KAFKA_INGESTION_WARNINGS,
     KAFKA_LOG_ENTRIES,
 } from '~/config/kafka-topics'
+import {
+    INGESTION_DOWNSTREAM_PRODUCER,
+    INGESTION_UPSTREAM_PRODUCER,
+    type ProducerName,
+} from '~/ingestion/common/producers'
 import { IngestionLane } from '~/ingestion/config'
 
 export type ErrorTrackingConsumerConfig = {
