+++ nodejs/src/ingestion/pipelines/heatmaps/consumer.test.ts
@@ -1,6 +1,6 @@
-import { ProducerName } from '~/common/outputs'
 import { KafkaProducerRegistry } from '~/common/outputs/kafka-producer-registry'
 import { CookielessManager } from '~/ingestion/common/cookieless/cookieless-manager'
+import { ProducerName } from '~/ingestion/common/producers'
 import { newScope } from '~/ingestion/common/scopes'
 import { IngestionOutputsConfig } from '~/ingestion/config'
 import { RedisPool } from '~/types'
