+++ nodejs/src/ingestion/pipelines/sessionreplay/outputs/producer-registry.ts
@@ -1,6 +1,5 @@
-import { INGESTION_DOWNSTREAM_PRODUCER } from '~/common/outputs'
 import { KafkaProducerRegistryBuilder } from '~/common/outputs/kafka-producer-registry-builder'
-import { INGESTION_DOWNSTREAM_PRODUCER_CONFIG_MAP } from '~/ingestion/common/config'
+import { INGESTION_DOWNSTREAM_PRODUCER, INGESTION_DOWNSTREAM_PRODUCER_CONFIG_MAP } from '~/ingestion/common/producers'
 import {
     INGESTION_SESSIONREPLAY_PRODUCER,
     INGESTION_SESSIONREPLAY_PRODUCER_CONFIG_MAP,
