+++ nodejs/tests/ingestion/pipelines/heatmaps/consumer-e2e.test.ts
@@ -1,11 +1,11 @@
 import { Message } from 'node-rdkafka'
 
-import { KafkaProducerRegistryComponent } from '~/common/outputs/registry'
+import { CookielessManagerComponent } from '~/ingestion/common/cookieless/cookieless-manager'
+import { KafkaProducerRegistryComponent } from '~/ingestion/common/producer-registry'
 import {
     getDefaultKafkaDownstreamProducerEnvConfig,
     getDefaultKafkaUpstreamProducerEnvConfig,
-} from '~/ingestion/common/config'
-import { CookielessManagerComponent } from '~/ingestion/common/cookieless/cookieless-manager'
+} from '~/ingestion/common/producers'
 import { Component, newScope } from '~/ingestion/common/scopes'
 import { getDefaultIngestionOutputsConfig } from '~/ingestion/config'
 import {
